Split out 2038 parsing tool from sample generation tool
The original tool had two key capabilities: the ability to parse
2038 streams and hand off their contents to libklvanc, and a
routine to generate a sample stream.  Move the sample stream
generation into a separate utility, which eliminates several
in-library dependencies from the parsing tool (i.e. klringbuffer,
ts_packetizer, klbitstream_readerwriter).  This will make it
easier to potentially move the parsing tool into a different
project or adapt for other purposes.
