+++ tools/klvanc_util.c
@@ -29,6 +29,7 @@ extern int parse_main(int argc, char *argv[]);
 extern int smpte2038_main(int argc, char *argv[]);
 extern int scte104_main(int argc, char *argv[]);
 extern int genscte104_main(int argc, char *argv[]);
+extern int gensmpte2038_main(int argc, char *argv[]);
 extern int eia708_main(int argc, char *argv[]);
 extern int smpte12_2_main(int argc, char *argv[]);
 extern int afd_main(int argc, char *argv[]);
@@ -47,6 +48,7 @@ int main(int argc, char *argv[])
 		{ "klvanc_scte104",		scte104_main, },
 		{ "klvanc_eia708",		eia708_main, },
 		{ "klvanc_genscte104",		genscte104_main, },
+		{ "klvanc_gensmpte2038",	gensmpte2038_main, },
 		{ "klvanc_smpte12_2",		smpte12_2_main, },
 		{ "klvanc_afd",			afd_main, },
 		{ 0, 0 },
