+++ tools/smpte2038.c
@@ -31,13 +31,9 @@
 #include <signal.h>
 #include <sys/time.h>
 #include <libklvanc/vanc.h>
-#include "klbitstream_readwriter.h"
 #include "udp.h"
 #include "url.h"
-#include "ts_packetizer.h"
-#include "klringbuffer.h"
 #include "pes_extractor.h"
-
 #include "version.h"
 #include "hexdump.h"
 
@@ -124,142 +120,6 @@ static pes_extractor_callback pes_cb(void *cb_context, uint8_t *buf, int byteCou
 	return 0;
 }
 
-/* Create a PES array containing 8 lines of VANC data.
- * Write it to disk (/tmp) and attempt to parse it to check the
- * parser is operating correctly.
- */
-static void smpte2038_generate_sample_708B_packet(struct app_context_s *ctx)
-{
-	/* STEP 1. Generate some useful VANC packed into a PES frame. */
-
-	/* This is a fully formed 708B VANC message. We'll bury
-	 * this inside a SMPTE2038 wrapper and prepare a TS packet
-	 * that could be used for sample/test data.
-	 */
-	unsigned short arr[] = {
-		0x000, 0x3ff, 0x3ff, 0x161, 0x101, 0x152, 0x296, 0x269,
-		0x152, 0x14f, 0x277, 0x2b8, 0x1ad, 0x272, 0x1f4, 0x2fc,
-		0x180, 0x180, 0x1fd, 0x180, 0x180, 0x2fa, 0x200, 0x200,
-		0x2fa, 0x200, 0x200, 0x2fa, 0x200, 0x200, 0x2fa, 0x200,
-		0x200, 0x2fa, 0x200, 0x200, 0x2fa, 0x200, 0x200, 0x2fa,
-		0x200, 0x200, 0x2fa, 0x200, 0x200, 0x2fa, 0x200, 0x200,
-		0x2fa, 0x200, 0x200, 0x2fa, 0x200, 0x200, 0x2fa, 0x200,
-		0x200, 0x2fa, 0x200, 0x200, 0x2fa, 0x200, 0x200, 0x2fa,
-		0x200, 0x200, 0x2fa, 0x200, 0x200, 0x2fa, 0x200, 0x200,
-		0x2fa, 0x200, 0x200, 0x173, 0x2d1, 0x1e0, 0x200, 0x200,
-		0x200, 0x200, 0x200, 0x200, 0x274, 0x2b8, 0x1ad, 0x194,
-		0x1b4, /* Checksum */
-	};
-
-	/* See smpte 2038-2008 - Page 5, Table 2 for description. */
-	struct klbs_context_s *bs  = klbs_alloc();
-
-	uint8_t buf[8192] = { 0 };
-	klbs_write_set_buffer(bs, buf, sizeof(buf));
-
-	/* PES Header - Bug: bitstream can't write 32bit values */
-	klbs_write_bits(bs, 1, 24);		/* packet_start_code_prefix */
-	klbs_write_bits(bs, 0xBD, 8);		/* stream_id */
-	klbs_write_bits(bs, 0, 16);		/* PES_packet_length */
-	klbs_write_bits(bs, 2, 2);		/* '10' fixed value */
-	klbs_write_bits(bs, 0, 2);		/* PES_scrambling_control (not scrambled) */
-	klbs_write_bits(bs, 0, 1);		/* PES_priority */
-	klbs_write_bits(bs, 1, 1);		/* data_alignment_indicator (aligned) */
-	klbs_write_bits(bs, 0, 1);		/* copyright (not-copyright) */
-	klbs_write_bits(bs, 0, 1);		/* original-or-copy (copy) */
-	klbs_write_bits(bs, 2, 2);		/* PTS_DTS_flags (PTS Present) */
-	klbs_write_bits(bs, 0, 1);		/* ESCR_flag (not present) */
-	klbs_write_bits(bs, 0, 1);		/* ES_RATE_flag (not present) */
-	klbs_write_bits(bs, 0, 1);		/* DSM_TRICK_MODE_flag (not present) */
-	klbs_write_bits(bs, 0, 1);		/* additional_copy_info_flag (not present) */
-	klbs_write_bits(bs, 0, 1);		/* PES_CRC_flag (not present) */
-	klbs_write_bits(bs, 0, 1);		/* PES_EXTENSION_flag (not present) */
-	klbs_write_bits(bs, 5, 8);		/* PES_HEADER_DATA_length */
-	klbs_write_bits(bs, 2, 4);		/* '0010' fixed value */
-
-	uint64_t pts = 0;
-	klbs_write_bits(bs, (pts >> 30), 3);	/* PTS[32:30] */
-	klbs_write_bits(bs, 1, 1);			/* marker_bit */
-	klbs_write_bits(bs, (pts >> 15) & 0xefff, 15);	/* PTS[29:15] */
-	klbs_write_bits(bs, 1, 1);			/* marker_bit */
-	klbs_write_bits(bs, (pts & 0xefff), 15);		/* PTS[14:0] */
-	klbs_write_bits(bs, 1, 1);			/* marker_bit */
-
-	int lineCount = 4;
-	for (int i = 0; i < lineCount; i++) {
-		/* VANC Payload */
-		klbs_write_bits(bs, 0, 6);	/* fixed value '000000' */
-		klbs_write_bits(bs, 0, 1);	/* c_not_y_channel_flag (HD luminance) */
-		klbs_write_bits(bs, 10 + i, 11);	/* line_number (9) */
-		klbs_write_bits(bs, 0, 12);	/* horizonal_offset (0 words from SAV) */
-		klbs_write_bits(bs, arr[3], 10);	/* DID */
-		klbs_write_bits(bs, arr[4], 10);	/* SDID */
-		klbs_write_bits(bs, arr[5], 10);	/* data_count */
-		for (int i = 6; i < (sizeof(arr) / sizeof(unsigned short)); i++) {
-			/* This data_count AND checksum */
-			klbs_write_bits(bs, arr[i], 10);
-		}
-
-		/* Byte alignment stuffing */
-		while (bs->reg_used > 0) {
-			klbs_write_bits(bs, 1, 1);	/* Stuffing byte */
-		}
-	}
-
-	/* Flush the remaining bits out into buffer, else we lose up to
-	 * the last 7 bits that are cached in the bitstream implementation.
-	 */
-	klbs_write_buffer_complete(bs);
-
-	/* Now updated the PES packet length */
-	int len = klbs_get_byte_count(bs) - 6;
-	klbs_get_buffer(bs)[4] = (len >> 8) & 0xff;
-	klbs_get_buffer(bs)[5] = len & 0xff;
-
-	/* STEP 2. Do something useful with the PES, now that its fully assembled. */
-
-	/* The PES is ready. Save a file copy. */
-	printf("%s() We've constructed a fake PES, here it is:\n", __func__);
-	hexdump(buf, klbs_get_byte_count(bs), 16);
-	klbs_save(bs, "/tmp/bitstream-scte2038-EIA708B.raw");
-
-	/* STEP 3. Maybe we should packetize the PES into TS packets. */
-
-	uint8_t section[8192];
-	int section_length = klbs_get_byte_count(bs);
-	memset(section, 0xff, sizeof(section));
-	memcpy(section, klbs_get_buffer(bs), section_length);
-
-	uint8_t *pkts = 0;
-	uint32_t packetCount = 0;
-	uint8_t cc = 0;
-	ts_packetizer(section, section_length, &pkts, &packetCount, 188, &cc, ctx->pid);
-	for (uint32_t i = 0; i < packetCount; i++) {
-		printf("%s() We've constructed some TS packets, here they are:\n", __func__);
-		hexdump(pkts + (i * 188), 188, 16);
-	}
-
-	/* Write all the TS packets to a temp file. */
-	FILE *fh = fopen("/tmp/bitstream-scte2038-EIA708B.ts", "wb");
-	if (fh) {
-		fwrite(pkts, packetCount, 188, fh);
-		fclose(fh);
-	}
-
-	/* STEP 4. Test the PES extractor, parse our TS packets, extract the PES
-	 * parse the SMPTE2038 lines. Expect out pes_cb to get called after the PES
-	 * has been fully assembled, our cb will parse the 2038 contained.
-	 */
-	struct pes_extractor_s *pe;
-	pe_alloc(&pe, ctx, (pes_extractor_callback)pes_cb, ctx->pid);
-	pe_push(pe, pkts, packetCount);
-	pe_free(&pe);
-
-	free(pkts); /* Results from the packetizer have to be caller freed. */
-
-	klbs_free(bs);
-}
-
 /* We're called with blocks of UDP data */
 static tsudp_receiver_callback udp_cb(void *userContext, uint8_t *buf, int byteCount)
 {
@@ -301,7 +161,6 @@ static int _usage(const char *progname, int status)
 		"    -i <udp url. Eg. udp://224.0.0.1:5000>\n"
 		"    -P <pid 0xNNNN> VANC PID to process (def: 0x%x)\n"
 		"    -v Increase verbose level\n"
-		"    -g generate sample SMPTE2038 stream and parse it.\n"
 		"    -t <vanc_types> enable VANC dumping (e.g. 'cea708,scte104')\n"
 		"       valid types are: all",
 	basename((char *)progname),
@@ -386,7 +245,6 @@ static int _main(int argc, char *argv[])
 {
 	int opt;
 	int exitStatus = 0;
-	int doGenerateSample = 0;
 	ctx->running = 1;
 	ctx->pid = DEFAULT_PID;
 	ctx->verbose = 0;
@@ -397,11 +255,8 @@ static int _main(int argc, char *argv[])
 	} inputType = IT_UDP;
 	static struct klvanc_callbacks_s callbacks;
 
-	while ((opt = getopt(argc, argv, "?ghi:P:vt:")) != -1) {
+	while ((opt = getopt(argc, argv, "?hi:P:vt:")) != -1) {
 		switch (opt) {
-		case 'g':
-			doGenerateSample = 1;
-			break;
 		case 'i':
 			ctx->input_url = optarg;
 			if (url_parse(ctx->input_url, &ctx->i_url) < 0) {
@@ -445,14 +300,6 @@ static int _main(int argc, char *argv[])
 		case 'h':
 			_usage(argv[0], 0);
 		}
-	}
-
-	if (doGenerateSample) {
-		/* Do this outside of the switch else -v becomes highly
-		 * command line position dependant.
-		 */
-		smpte2038_generate_sample_708B_packet(ctx);
-		exit(0);
 	}
 
 	if (ctx->input_url == NULL) {
